Fixes #47 (allow arg names in service function definitions)
