+++ aqua-src/error.aqua
@@ -1,5 +1,7 @@
 service Demo("demo"):
-	get4: -> u64
+	get4() -> u64
+	get5(arg: u32)
+	get6(a: string) -> bool
 
 func one() -> u64:
     variable <- Demo.get4()
