+++ semantics/src/main/scala/aqua/semantics/rules/abilities/AbilitiesInterpreter.scala
@@ -1,7 +1,7 @@
 package aqua.semantics.rules.abilities
 
 import aqua.semantics.rules.{ReportError, StackInterpreter}
-import aqua.parser.lexer.{Ability, Name, Token, Value}
+import aqua.parser.lexer.{Name, Value}
 import aqua.types.ArrowType
 import cats.data.{NonEmptyList, NonEmptyMap, State}
 import cats.~>
@@ -12,7 +12,7 @@ import monocle.macros.GenLens
 class AbilitiesInterpreter[F[_], X](implicit
   lens: Lens[X, AbilitiesState[F]],
   error: ReportError[F, X]
-) extends StackInterpreter[F, X, AbilitiesState[F], AbilityStackFrame[F]](
+) extends StackInterpreter[F, X, AbilitiesState[F], AbilitiesState.Frame[F]](
       GenLens[AbilitiesState[F]](_.stack)
     ) with (AbilityOp[F, *] ~> State[X, *]) {
 
@@ -22,7 +22,7 @@ class AbilitiesInterpreter[F[_], X](implicit
   override def apply[A](fa: AbilityOp[F, A]): State[X, A] =
     (fa match {
       case bs: BeginScope[F] =>
-        beginScope(AbilityStackFrame[F](bs.token))
+        beginScope(AbilitiesState.Frame[F](bs.token))
 
       case EndScope() =>
         endScope
@@ -112,26 +112,3 @@ class AbilitiesInterpreter[F[_], X](implicit
 
     }).asInstanceOf[State[X, A]]
 }
-
-case class AbilitiesState[F[_]](
-  stack: List[AbilityStackFrame[F]] = Nil,
-  services: Map[String, NonEmptyMap[String, ArrowType]] = Map.empty,
-  rootServiceIds: Map[String, Value[F]] = Map.empty[String, Value[F]],
-  definitions: Map[String, Ability[F]] = Map.empty[String, Ability[F]]
-) {
-
-  def purgeArrows: Option[(NonEmptyList[(Name[F], ArrowType)], AbilitiesState[F])] =
-    stack match {
-      case sc :: tail =>
-        NonEmptyList
-          .fromList(sc.arrows.values.toList)
-          .map(_ -> copy[F](sc.copy(arrows = Map.empty) :: tail))
-      case _ => None
-    }
-}
-
-case class AbilityStackFrame[F[_]](
-  token: Token[F],
-  arrows: Map[String, (Name[F], ArrowType)] = Map.empty[String, (Name[F], ArrowType)],
-  serviceIds: Map[String, Value[F]] = Map.empty[String, Value[F]]
-)
