+++ semantics/src/main/scala/aqua/semantics/rules/names/NamesState.scala
@@ -1,0 +1,43 @@
+package aqua.semantics.rules.names
+
+import aqua.parser.lexer.{Name, Token}
+import aqua.types.{ArrowType, Type}
+import cats.kernel.Monoid
+
+case class NamesState[F[_]](
+  stack: List[NamesState.Frame[F]] = Nil,
+  rootArrows: Map[String, ArrowType] = Map.empty,
+  definitions: Map[String, Name[F]] = Map.empty[String, Name[F]]
+) {
+
+  def allNames: LazyList[String] =
+    LazyList.from(stack).flatMap(s => s.names.keys ++ s.arrows.keys).appendedAll(rootArrows.keys)
+
+  def allArrows: LazyList[String] =
+    LazyList.from(stack).flatMap(_.arrows.keys).appendedAll(rootArrows.keys)
+}
+
+object NamesState {
+
+  case class Frame[F[_]](
+    token: Token[F],
+    names: Map[String, Type] = Map.empty,
+    arrows: Map[String, ArrowType] = Map.empty
+  ) {
+    def addName(n: String, t: Type): NamesState.Frame[F] = copy[F](names = names.updated(n, t))
+
+    def addArrow(n: String, g: ArrowType): NamesState.Frame[F] =
+      copy[F](arrows = arrows.updated(n, g))
+  }
+
+  implicit def namesStateMonoid[F[_]]: Monoid[NamesState[F]] = new Monoid[NamesState[F]] {
+    override def empty: NamesState[F] = NamesState[F]()
+
+    override def combine(x: NamesState[F], y: NamesState[F]): NamesState[F] =
+      NamesState(
+        stack = Nil,
+        rootArrows = x.rootArrows ++ y.rootArrows,
+        definitions = x.definitions ++ y.definitions
+      )
+  }
+}
